requirements.txt
-e git://github.com/django/django.git@f43da05cc5c76102c04fd2e76b4bada880265dbe#egg=Django
django-tastypie==0.13.3
gunicorn==19.4.5
python-dateutil==2.5.2
python-mimeparse==1.5.1
six==1.10.0
